+++ R/Confirmatory-methods.R
@@ -114,7 +114,7 @@ setMethod(
     f = "residuals",
     signature = signature(object = 'ConfirmatoryClass'),
     definition = function(object, restype = 'LD', digits = 3, df.p = FALSE, full.scores = FALSE,
-                          printvalue = NULL, verbose = TRUE, ...)
+                          printvalue = NULL, tables = FALSE, verbose = TRUE, ...)
     {
         K <- object@K
         data <- object@data
@@ -135,6 +135,8 @@ setMethod(
         df <- (object@K - 1) %o% (object@K - 1)
         diag(df) <- NA
         colnames(df) <- rownames(df) <- colnames(res)
+        itemnames <- colnames(data)
+        listtabs <- list()
         if(restype == 'LD'){
             for(i in 1:J){
                 for(j in 1:J){
@@ -151,9 +153,14 @@ setMethod(
                         res[j,i] <- sum(((tab - Etab)^2)/Etab) * sign(s)
                         res[i,j] <- sign(res[j,i]) * sqrt( abs(res[j,i]) / (N*min(c(K[i],K[j]) - 1L)))
                         df[i,j] <- pchisq(abs(res[j,i]), df=df[j,i], lower.tail=FALSE)
+                        if(tables){
+                            tmp <- paste0(itemnames[i], '_', itemnames[j])
+                            listtabs[[tmp]] <- list(Obs=tab, Exp=Etab, std_res=(tab-Etab)/sqrt(Etab))
+                        }
                     }
                 }
             }
+            if(tables) return(listtabs)
             if(df.p){
                 cat("Degrees of freedom (lower triangle) and p-values:\n\n")
                 print(round(df, digits))